Fixed some src modules and updated to devel version.

+++ src/modules.c
@@ -539,13 +539,14 @@ do_modlist(struct Client *source_p, const char *pattern)
                 sendto_one(source_p, form_str(RPL_MODLIST),
                            me.name, source_p->name,
                            modlist[i]->name,
-                           modlist[i]->address,
+                           (unsigned long)(uintptr_t)modlist[i]->address,
                            modlist[i]->version, modlist[i]->core ? "(core)" : "");
             }
         } else {
             sendto_one(source_p, form_str(RPL_MODLIST),
                        me.name, source_p->name, modlist[i]->name,
-                       modlist[i]->address, modlist[i]->version,
+                       (unsigned long)(uintptr_t)modlist[i]->address,
+                       modlist[i]->version,
                        modlist[i]->core ? "(core)" : "");
         }
     }
